-fix tests -correct naming -remove dead code
